refactor(index): split the query_api spine into search/ai_lifecycle/gc submodules
query_api/mod.rs had drifted back into a leftover-methods spine. Move the
cohesive method groups into sibling submodules at the same depth (preserving
pub(super) reach):

- search.rs: search/search_explain/search_with_graph_meta, search_hash_baseline,
  docs_for_symbol/docs_for_selected_symbol, plus the SearchRequest builder.
- ai_lifecycle.rs: the on-device embedding-model lifecycle (status, list/install
  model, reconcile*, current_embedding_count).
- gc.rs: gc/prune_to_live and the GcReport result type they own (relocated from
  index/mod.rs to sit with its only producers).

mod.rs drops from 733 to 456 lines; no behavior change.
