+++ crates/rag-rat-core/src/index/query_api/ai_lifecycle.rs
@@ -1,0 +1,53 @@
+//! On-device embedding-model lifecycle on `IndexDatabase`: status, model install/list, and the
+//! reconcile (re-embed) entry points. Thin pass-throughs to `crate::index::ai`; kept together so
+//! the model-management surface reads as one unit.
+
+use super::*;
+
+impl IndexDatabase {
+    pub fn local_ai_status(&self) -> anyhow::Result<LocalAiStatus> {
+        ai::status(self.storage.connection())
+    }
+
+    pub fn list_models(&self) -> anyhow::Result<Vec<ModelInfo>> {
+        ai::models(self.storage.connection())
+    }
+
+    pub fn install_model(&self, model_id: &str) -> anyhow::Result<ModelInfo> {
+        ai::install_model(self.storage.connection(), model_id)
+    }
+
+    pub fn reconcile(
+        &self,
+        limit: Option<u32>,
+        batch_size: Option<u32>,
+    ) -> anyhow::Result<ReconcileReport> {
+        ai::reconcile(self.storage.connection(), limit, batch_size)
+    }
+
+    pub fn reconcile_plan(&self) -> anyhow::Result<ReconcilePlan> {
+        ai::reconcile_plan(self.storage.connection())
+    }
+
+    pub fn reconcile_with_progress(
+        &self,
+        limit: Option<u32>,
+        batch_size: Option<u32>,
+        force: bool,
+        progress: impl FnMut(ai::ReconcileProgress),
+    ) -> anyhow::Result<ReconcileReport> {
+        ai::reconcile_with_progress(self.storage.connection(), limit, batch_size, force, progress)
+    }
+
+    pub fn reconcile_with_options_progress(
+        &self,
+        options: ai::ReconcileOptions,
+        progress: impl FnMut(ai::ReconcileProgress),
+    ) -> anyhow::Result<ReconcileReport> {
+        ai::reconcile_with_options_progress(self.storage.connection(), options, progress)
+    }
+
+    pub fn current_embedding_count(&self, model_id: &str) -> anyhow::Result<u64> {
+        ai::current_embedding_count(self.storage.connection(), model_id)
+    }
+}
