+++ crates/rag-rat-core/src/index/query_api/gc.rs
@@ -1,0 +1,141 @@
+//! Index garbage collection on `IndexDatabase`: prune file/chunk/embedding/symbol/edge rows for
+//! git contexts that are no longer live, plus the [`GcReport`] it returns.
+
+use serde::Serialize;
+
+use super::*;
+
+#[derive(Debug, Clone, Serialize)]
+pub struct GcReport {
+    pub files_pruned: u64,
+    pub chunks_pruned: u64,
+    pub files_remaining: u64,
+    pub chunks_remaining: u64,
+    /// True when no live context could be determined and pruning was skipped (nothing deleted).
+    pub skipped: bool,
+}
+
+impl IndexDatabase {
+    /// Garbage-collect index rows for git contexts that are no longer live. Keeps the active
+    /// commit and overlay of every worktree reported by `git worktree list` (plus this
+    /// connection's active context) and prunes file/chunk/embedding/symbol/edge rows for any
+    /// other commit. Never prunes when no live context can be determined (non-git, git error).
+    pub fn gc(&self) -> anyhow::Result<GcReport> {
+        let mut live_commits = Vec::new();
+        let mut live_worktrees = Vec::new();
+        if let Some(root) = self.storage.source_root() {
+            let (commits, worktrees) = live_worktree_contexts(root);
+            live_commits.extend(commits);
+            live_worktrees.extend(worktrees);
+        }
+        // Always keep this connection's active context, even if git enumeration missed it.
+        if !self.active_commit_sha.is_empty() {
+            live_commits.push(self.active_commit_sha.clone());
+        }
+        if !self.active_worktree_id.is_empty() {
+            live_worktrees.push(self.active_worktree_id.clone());
+        }
+        live_commits.sort();
+        live_commits.dedup();
+        live_worktrees.sort();
+        live_worktrees.dedup();
+        self.prune_to_live(&live_commits, &live_worktrees)
+    }
+
+    /// Prune file rows (and their derived rows) whose `commit_sha` and `worktree_id` are both
+    /// outside the live sets. Refuses to prune when both live sets are empty, so a missing
+    /// live set never wipes the index. `parser_failures` are keyed by path (shared across
+    /// commits) and are regenerated on the next index, so they are not preserved per-commit.
+    pub fn prune_to_live(
+        &self,
+        live_commits: &[String],
+        live_worktrees: &[String],
+    ) -> anyhow::Result<GcReport> {
+        let conn = self.storage.connection();
+        let files_before = table_row_count(conn, "files")?;
+        let chunks_before = table_row_count(conn, "chunks")?;
+        if live_commits.is_empty() && live_worktrees.is_empty() {
+            return Ok(GcReport {
+                files_pruned: 0,
+                chunks_pruned: 0,
+                files_remaining: files_before,
+                chunks_remaining: chunks_before,
+                skipped: true,
+            });
+        }
+        conn.execute_batch(
+            "
+            CREATE TEMP TABLE IF NOT EXISTS gc_live_commits(sha TEXT PRIMARY KEY);
+            DELETE FROM temp.gc_live_commits;
+            CREATE TEMP TABLE IF NOT EXISTS gc_live_worktrees(id TEXT PRIMARY KEY);
+            DELETE FROM temp.gc_live_worktrees;
+            CREATE TEMP TABLE IF NOT EXISTS staged_file_ids(id INTEGER PRIMARY KEY);
+            DELETE FROM temp.staged_file_ids;
+            ",
+        )?;
+        {
+            let mut stmt =
+                conn.prepare("INSERT OR IGNORE INTO temp.gc_live_commits(sha) VALUES (?1)")?;
+            for sha in live_commits {
+                stmt.execute([sha])?;
+            }
+        }
+        {
+            let mut stmt =
+                conn.prepare("INSERT OR IGNORE INTO temp.gc_live_worktrees(id) VALUES (?1)")?;
+            for id in live_worktrees {
+                stmt.execute([id])?;
+            }
+        }
+        // A file survives if its commit is live OR its worktree overlay is live. Empty-string
+        // keys never appear in the live sets, so unkeyed rows are pruned.
+        conn.execute(
+            "
+            INSERT OR IGNORE INTO temp.staged_file_ids(id)
+            SELECT id FROM main.files
+            WHERE commit_sha NOT IN (SELECT sha FROM temp.gc_live_commits)
+              AND worktree_id NOT IN (SELECT id FROM temp.gc_live_worktrees)
+            ",
+            [],
+        )?;
+        self.delete_staged_files_cascade()?;
+        conn.execute_batch("DELETE FROM temp.staged_file_ids;")?;
+        // `edge_oracle` verdicts cascade away with their edges via the FK ON DELETE CASCADE (fired
+        // by the cascade above, with `PRAGMA foreign_keys=ON`). `oracle_runs`, however, is keyed by
+        // `(commit_sha, worktree_id)` directly — nothing cascades it — so a dead checkout's run
+        // rows would survive the file pruning. Prune them with the SAME live sets, so a run and the
+        // edges it produced are dropped together.
+        oracle::prune_oracle_runs_outside_scope(conn, live_commits, live_worktrees)?;
+        // Dictionary hygiene (#79): drop `edge_strings` values no edge references any more. The
+        // dictionary has NO FKs by design (see the schema comment), so orphans accumulate as
+        // edges are pruned; the vocabulary is small, but gc is the natural rate-limited home for
+        // the sweep. Every referencing column must appear here — a missed column would null its
+        // strings out from under live edges.
+        conn.execute(
+            "
+            DELETE FROM main.edge_strings
+            WHERE id NOT IN (
+                SELECT from_name_id FROM main.edges_data WHERE from_name_id IS NOT NULL
+                UNION SELECT to_name_id FROM main.edges_data
+                UNION SELECT target_qualified_name_id FROM main.edges_data
+                    WHERE target_qualified_name_id IS NOT NULL
+                UNION SELECT receiver_hint_id FROM main.edges_data
+                    WHERE receiver_hint_id IS NOT NULL
+                UNION SELECT resolution_id FROM main.edges_data
+                UNION SELECT edge_kind_id FROM main.edges_data
+                UNION SELECT confidence_id FROM main.edges_data
+            )
+            ",
+            [],
+        )?;
+        let files_remaining = table_row_count(conn, "files")?;
+        let chunks_remaining = table_row_count(conn, "chunks")?;
+        Ok(GcReport {
+            files_pruned: files_before.saturating_sub(files_remaining),
+            chunks_pruned: chunks_before.saturating_sub(chunks_remaining),
+            files_remaining,
+            chunks_remaining,
+            skipped: false,
+        })
+    }
+}
